Fix issue with Figaro gem. Movie service passes initial tests
