+++ .rspec
@@ -1,3 +1,3 @@
 --require spec_helper
 --color
---format NyanCatMusicFormatter
+--format NyanCatFormatter
