+++ Gemfile
@@ -16,6 +16,7 @@ gem 'coffee-rails', '~> 4.2'
 gem 'bootsnap'
 gem 'jbuilder', '~> 2.5'
 gem 'bcrypt', '~> 3.1.7'
+gem 'figaro'
 
 
 group :development, :test do
@@ -47,7 +48,7 @@ end
 # end
 
 gem 'faraday'
-gem 'figaro'
+
 gem 'json'
 gem 'sinatra'
