+++ app/poros/movie_info.rb
@@ -10,4 +10,6 @@ end
 # response[0][:title]
 # response[0][:vote_average]
 
-# movie_stuff = MovieInfo.new 
+movie_stuff = MovieInfo.new 
+
+movie_stuff.popular_movies
