--- app/service/movie_service.rb
@@ -1,16 +1,0 @@
-require 'faraday'
-require 'json'
-
-class MovieService
-  def get_movie_data
-    response1 = Faraday.get("https://api.themoviedb.org/3/movie/popular?api_key=ENV['movie_api_key']&language=en-US&page=1")
-    response2 = Faraday.get("https://api.themoviedb.org/3/movie/popular?api_key=ENV['movie_api_key']&language=en-US&page=2")
-    json1 = JSON.parse(response1.body, symbolize_names: true)
-    json2 = JSON.parse(response2.body, symbolize_names: true)
-    response = json1[:results] + json2[:results] # 40 movies returned
-  end
-end
-
-test = MovieService.new 
-
-test.get_movie_data
