+++ spec/models/user_spec.rb
@@ -14,5 +14,6 @@ RSpec.describe User do
         it {should have_many(:followees).through(:followed_users)}
         it {should have_many :following_users}
         it {should have_many(:followers).through(:following_users)}
+        require 'pry'; binding.pry
     end
 end
